requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
Pillow==10.0.0
psycopg2-binary==2.9.7
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
yarl==1.9.2
gspread~=5.10.0
oauth2client~=4.1.3
django-import-export==3.2.0